backend/pyproject.toml
[project]
name = "saymed_pain_bprc_backend"
version = "0.1.0"
description = "FastAPI backend for SayMed Pain Edition (BPRC)"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "psycopg2-binary>=2.9",
  "pydantic[email]>=2.7",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]==1.7.4",
  "bcrypt==3.2.2",
  "python-multipart>=0.0.9",
  "reportlab>=4.0",
  "openai>=1.0.0",
  "python-dotenv>=1.0.0",
]
[tool.uvicorn]
factory = false
